Fix comments of teacher
The correction was:
The whole entity class was public. This is a bad design, member variables should be, usually, private or protected.
Function Player::getItems does not makes sense. Why you pass the player as parameter? Furthermore, the std::vector should be returned by using a const reference, otherwise, the whole vector will be copied during the function call and that's bad in terms of performance but also in terms of memory fragmentation.
Good use of forward declarations.
Good use of virtual destructor in Entity class.
Good use of const and references in general.

What i've done to fix the issues:
- Delete entity class not used
- Fix the function getItems